Refactoring Image Service

+++ src/main/java/com/java/profileservice/contoller/ImageController.java
@@ -2,7 +2,11 @@ package com.java.profileservice.contoller;
 
 
 import com.java.profileservice.config.ApiResponse;
+import com.java.profileservice.exceptions.EntityNotExistsException;
+import com.java.profileservice.model.Image;
+import com.java.profileservice.model.Profile;
 import com.java.profileservice.service.ImageService;
+import com.java.profileservice.service.ProfileService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.slf4j.Logger;
@@ -10,6 +14,10 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.security.Principal;
+import java.util.Objects;
+import java.util.Optional;
 
 /**
  * Image controller for Adopt a pet project
@@ -28,6 +36,9 @@ public class ImageController {
     @Autowired
     private ImageService imageService;
 
+    @Autowired
+    private ProfileService profileService;
+
     /**
      * 1) Delete image route uses to delete image
      *
@@ -38,7 +49,16 @@ public class ImageController {
             notes = "${operation15.description}",
             value = "${operation15.value}"
     )
-    public ApiResponse deleteImage(@PathVariable(name = "id") Long id) throws Exception {
+    public ApiResponse deleteImage(@PathVariable(name = "id") Long id, Principal principal) throws Exception {
+
+        if (Objects.isNull(principal)) {
+            throw new Exception("Authorization error, you have to be sign in first!");
+        }
+        Image image = imageService.getImageById(id);
+        if(!principal.getName().equalsIgnoreCase(image.getProfile().getUserName())){
+            throw new Exception("Access denied, you cannot delete this image!");
+        }
+
         LOG.info("Starting deleting image.");
         imageService.deleteImageById(id);
         LOG.info("Image has deleted.");
@@ -57,7 +77,17 @@ public class ImageController {
      * @param multipartFiles, profileId
      */
     public ApiResponse saveImage(@RequestParam(value = "files", required = true) MultipartFile[] multipartFiles,
-                                 @RequestParam(value = "profileId", required = true) Long profileId){
+                                 @RequestParam(value = "profileId", required = true) Long profileId,
+                                 Principal principal) throws Exception {
+
+        if (Objects.isNull(principal)) {
+            throw new Exception("Authorization error, you have to be sign in first!");
+        }
+        Profile profile = Optional.ofNullable(profileService.getProfileById(profileId)).orElseThrow(
+                () -> new EntityNotExistsException("Profile with id: " + profileId + " does not exists."));
+        if(!principal.getName().equalsIgnoreCase(profile.getUserName())){
+            throw new Exception("You cannot add image for this profile!");
+        }
         LOG.info("Starting saving image.");
         imageService.uploadImages(multipartFiles, profileId);
         LOG.info("Image has saved.");
